Move GroupDTO out of controller + validate permission
